synchronization, mutex locks and semaphores
Last week we are introduced to some concepts of synchronization in operating systems. The main challenge is to ensure data integrity when multiple threads/tasks are running concurrently.

There are several classic situations in regard of this endeavor. Among them one particular question caught my attention. It is the dining philosophers problem.